Check for vulnerable dependencies on CI
Currently the build only fails for vulnerability scores over 8,
but ideally this theshold should be lower (e.g. 4 for medium or above).

At the time of writing it's not clear how we can fix a particular CVE
(relating to org.springframework.security:spring-security-rsa:1.0.7.RELEASE)
so this will need to be revisited.